#82 <fix>: SecurityConfig에 WebMvcConfigurer을 implements 하면 controller 테스트코드가 실패하던 현상 수정
